Modify regular expressions
