--- api/pruebas.js
@@ -1,6 +1,0 @@
-function FindNeedle (h, n) {
-  //your code here   
-}
-
-console.log(FindNeedle('react-redux', 'redux'))
-console.log(FindNeedle('pinky', 'puntual'))
